allinbloomfastapi/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.13.3
pydantic==2.10.6
pydantic-settings==2.7.1
psycopg[binary]==3.2.5
python-jose[cryptography]==3.3.0
httpx==0.27.2
stripe==11.3.0
python-multipart==0.0.12
cuid2==2.0.1
google-auth==2.35.0
